AMBARI-26529 Fix Kerberos admin session ajax failure and avoid unnecessary KERBEROS service deletion

+++ ambari-web/app/controllers/main/admin/kerberos/wizard_controller.js
@@ -223,16 +223,19 @@ App.KerberosWizardController = App.WizardController.extend(App.InstallComponent,
     if (cachedService) {
       App.cache.services.removeObject(cachedService);
     }
-    if (modelService.get('isLoaded')) {
+    if (modelService && modelService.get('isLoaded')) {
       App.serviceMapper.deleteRecord(modelService);
+      return App.ajax.send({
+        name: 'common.delete.service',
+        sender: this,
+        data: {
+          serviceName: serviceName,
+          clusterName: App.get('clusterName')
+        }
+      });
+    } else {
+      return $.Deferred().resolve().promise();
     }
-    return App.ajax.send({
-      name: 'common.delete.service',
-      sender: this,
-      data: {
-        serviceName: serviceName
-      }
-    });
   },
 
   /**
